requirements.txt
flask
flask-restful
flask-sqlalchemy
pymysql
flask-migrate
flask-script
flask-jwt-extended
password_strength
requests
mock
flask-cors
mailjet_rest
